requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
python-dotenv==1.0.1
Pillow==10.2.0
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.1
django-filter==23.5
reportlab==4.1.0
openpyxl==3.1.2
requests==2.31.0 